backend/requirements.txt

flask>=2.3,<3.0

gunicorn>=21.0,<22.0

psycopg2-binary>=2.9,<3.0


